+Se pone el atributo de la clase Budget que representa el balance como int.
git-svn-id: http://svn2.xp-dev.com/svn/tp-tecnicas/trunk/tp-tecnicas@73 c5b30475-30d6-408d-831a-0400e7537d55

+++ src/model/game/Budget.java
@@ -1,21 +1,21 @@
 package model.game;
 
 public class Budget {
-	private float balance;
+	private int balance;
 
-	public Budget(float initialBalance) {
+	public Budget(int initialBalance) {
 		this.balance = initialBalance;
 	}
 	
-	public float getBalance() {
+	public int getBalance() {
 		return balance;
 	}
 	
-	public void increment(float amount) {
+	public void increment(int amount) {
 		this.balance += amount;
 	}
 	
-	public void decrement(float amount) {
+	public void decrement(int amount) {
 		this.balance -= amount;
 	}
